in progress
Broken right now because the way I was thinking of doing it with :loc
being an extra property on any kind of object won't work. Numbers are
usually primitives that can't have extra properties. Even when using a
wrapper object, it's difficult to clone the wrapped number to adjust the
properties; would have to do it case by case for every wrapper type. So
I think having a notion of syntax object that wraps data as in Racket
makes more sense.